backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
websockets>=12.0
google-cloud-speech>=2.21.0
google-generativeai>=0.3.0
google-genai>=1.0.0
elevenlabs>=1.0.0
google-api-python-client>=2.100.0
google-auth>=2.45.0
google-auth-oauthlib>=1.2.0
firebase-admin>=6.0.0
fitbit>=0.3.1
mem0ai>=0.1.0
